backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Video Processing
moviepy==1.0.3
opencv-python==4.8.1.78
Pillow==10.1.0
ffmpeg-python==0.2.0

# Audio Transcription
openai-whisper==20231117
# Use versions compatible with Windows/Python 3.12
torch==2.4.1
torchaudio==2.4.1

# Database and Storage
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
python-dotenv==1.0.0
boto3==1.34.0  # AWS S3 support

# Task Queue for async processing
celery==5.3.4
redis==5.0.1

# Authentication
firebase-admin==6.3.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uuid==1.30